The compiler had been talking for three days and nobody was listening
tools/warning_audit.py compiles all four builds clean with -Wall and holds
every warning against tools/warnings.txt, keyed by function rather than
line so the ledger survives ordinary edits. 819 warnings, 198 sites.

Its first run found that make -f Makefile.win had not compiled since
2026-08-29 -- 62 commits, on a C++17 construct mingw g++ 10 rejects at
its default -std, with -fpermissive downgrading the error and -w
swallowing the message. Three work log items listed "Windows builds"
among their nets in that window. Fixed with -std=gnu++17 and the three
-Wno-* flags that are all -w was ever doing; that build now reports zero
warnings, and win-tests.sh passes again.

Then the twelve warnings the compiler was right about: a != comparing a
pointer to itself that leaked every -YIC string, two provable overflows
(the JPL Horizons URL and a format string), six varargs type mismatches
including the first two diagnostics wdriver.cpp has ever had, a DrawCircle
that only looked guarded, and three sites of correct code that reads wrong.
-Wformat=, -Wformat-overflow=, -Wtautological-compare, -Wdangling-else,
-Wmisleading-indentation and -Wlogical-not-parentheses are now all empty.

Nets: suite 3526/0; chart matrix 0 of 6,936; switch matrix 2 of 75,635,
being exactly the leak that stopped happening; a multi-wheel bitmap
differential identical over four modes, proven pairwise distinct and
sabotage-proven; settings round trip; six audits and three generated
tables clean; win-tests 2 scenarios; all four builds compile, checked
rather than assumed.

+++ general.cpp
@@ -2281,7 +2281,7 @@ char *SzColor(KI ki)
   else {
     // Compose RGB color value.
     neg(ki);
-    sprintf2(S(szCol), "#%06x", Rgb(RgbB(ki), RgbG(ki), RgbR(ki)));
+    sprintf2(S(szCol), "#%06x", (uint)Rgb(RgbB(ki), RgbG(ki), RgbR(ki)));
   }
   return szCol;
 }
@@ -2299,7 +2299,7 @@ char *SzColor2(KI ki)
   else {
     // Compose RGB color value.
     neg(ki);
-    sprintf2(S(szCol), "#%06x", Rgb(RgbB(ki), RgbG(ki), RgbR(ki)));
+    sprintf2(S(szCol), "#%06x", (uint)Rgb(RgbB(ki), RgbG(ki), RgbR(ki)));
   }
   return szCol;
 }
@@ -2330,7 +2330,7 @@ CONST char *SzColorHTML(KI ki)
     kv = -ki;
 
   // Otherwise compose the RGB value for this color.
-  sprintf2(S(szCol), "#%06x", Rgb(RgbB(kv), RgbG(kv), RgbR(kv)));
+  sprintf2(S(szCol), "#%06x", (uint)Rgb(RgbB(kv), RgbG(kv), RgbR(kv)));
   return szCol;
 }
